Extract `auth` module
